requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
coverage==6.0.2
Django==3.2.8
django-filter==21.1
djangorestframework==3.12.4
idna==3.3
mock==4.0.3
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7
xlwt==1.3.0
